♻️  macros: Split `service` tests into separate files
